requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-huggingface
langchain-community
faiss-cpu
transformers
sentence-transformers
torch
python-dotenv
tiktoken
